+++ src/index.web.js
@@ -20,11 +20,7 @@ export { default as delay } from './utils/delay';
 export { default as wrapProps } from './utils/wrapProps';
 export { default as backToHome } from './utils/backToHome';
 export { default as htmlFilter } from './utils/htmlFilter';
-import * as LocalStorage from './utils/storage/localStorage';
-import * as SessionStorage from './utils/storage/sessionStorage';
 import * as MobileDetect from './utils/mobileDetect';
 export {
-    LocalStorage,
-    SessionStorage,
     MobileDetect
 };